Create folder from productModel.js and refactoring the code in PRO folder
