+++ PRO/controllers/productController.js
@@ -1,18 +1,19 @@
-const fs = require("fs");
+// const fs = require("fs");
+const Product = require('./../models/productModel');
 
-const products = JSON.parse(fs.readFileSync(`${__dirname}/../dev-data/data/products-simple.json`));
+// const products = JSON.parse(fs.readFileSync(`${__dirname}/../dev-data/data/products-simple.json`));
 
-exports.checkID = (req , res , next , val) => {
-  console.log(`Product id is ${val}`);
+// exports.checkID = (req , res , next , val) => {
+//   console.log(`Product id is ${val}`);
   
-  if (req.params.id * 1 > products.length) {
-    return res.status(404).json({
-      status: "fail",
-      message: "Invalid ID",
-    });
-  }
-  next();
-}
+//   if (req.params.id * 1 > products.length) {
+//     return res.status(404).json({
+//       status: "fail",
+//       message: "Invalid ID",
+//     });
+//   }
+//   next();
+// }
 
 exports.checkBody = (req , res , next) => {
    if(!req.body.name || !req.body.price) {
@@ -30,41 +31,32 @@ exports.getAllProducts = (req , res) => {
   res.status(200).json({
     status: "successfully added the products in list",
     requestedAt: req.requestTime ,
-    result: products.length ,
-    data: {
-      products,
-    },
+    // result: products.length ,
+    // data: {
+    //   products,
+    // },
   });
 };
 
 exports.getProduct = (req , res) => {
   console.log(req.params);
   const id = req.params.id * 1;
-  const product = products.find((el) => el.id === id);
-  res.status(200).json({
-    status: "success",
-    data: {
-      product,
-    },
-  });
+  // const product = products.find((el) => el.id === id);
+  // res.status(200).json({
+  //   status: "success",
+  //   data: {
+  //     product,
+  //   },
+  // });
 };
 
 exports.createProduct = (req , res) => {
-  const newId = products[products.length - 1].id + 1;
-  const newProduct = Object.assign({ id: newId }, req.body);
-
-  products.push(newProduct);
-
-  fs.writeFile(`${__dirname}/dev-data/data/products-simple.json`, JSON.stringify(products) , (err) => {
-      // The HTTP  status 201 Created success status response code indicates that the request has succeeded and has led to the creation of a resource.
-      res.status(201).json({
-        status: "success",
-        data: {
-          product: newProduct,
-        },
-      });
+  res.status(201).json({
+    status: "success",
+    data: {
+      product: newProduct,
     }
-  );
+  });
 };
 
 exports.updateProduct = (req , res) => {
