+++ PRO/routes/productRoutes.js
@@ -3,7 +3,7 @@ const productController = require('./../controllers/productController');
 
 const router = express.Router();
 
-router.param('id' , productController.checkID);
+// router.param('id' , productController.checkID);
 
 router
   .route('/')
